requirements.txt
alembic==0.9.9
certifi==2018.4.16
chardet==3.0.4
click==6.7
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
markdown2==2.3.5
MarkupSafe==1.0
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.8
urllib3==1.23
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
